@mszarma 2021-07-07
> did you benchmark how fast is this implementation?

I benchmarked second version on synthetic bench, here are some examples of the results:

OLD = remove_edges + neighbor_sampling call
NEW = neighbor_sampling with excluded edges call
==================================================
old=0.0427086 , new=0.0160954 , diff=0.0266132 seconds, diff=62.31 percentage , faster

GRAPH num_nodes: 1019999  num_edges: 29855 
sampled_nodes_amount: 174  fanout: 3 excluded edges amount: 168 
SUBGRAPH num_nodes: 1019999  num_edges: 231 
==================================================
==================================================
old=12.5179353 , new=6.0757782 , diff=6.4421570 seconds, diff=51.46 percentage , faster

GRAPH num_nodes: 80002000  num_edges: 239989054 
sampled_nodes_amount: 689396  fanout: 6 excluded edges amount: 1990128 
SUBGRAPH num_nodes: 80002000  num_edges: 4136376
==================================================
==================================================
old=8.8966670 , new=6.6231253 , diff=2.2735417 seconds, diff=25.55 percentage , faster

GRAPH num_nodes: 80020000  num_edges: 239989189 
sampled_nodes_amount: 94441  fanout: 6 excluded edges amount: 76725 
SUBGRAPH num_nodes: 80020000  num_edges: 566646
==================================================
==================================================
old=0.0245998 , new=0.0203795 , diff=0.0042202 seconds, diff=17.16 percentage , faster

GRAPH num_nodes: 299992  num_edges: 149993 
sampled_nodes_amount: 8109  fanout: 178 excluded edges amount: 15545 
SUBGRAPH num_nodes: 299992  num_edges: 5420
==================================================
==================================================
old=0.0080178 , new=0.0049868 , diff=0.0030310 seconds, diff=37.80 percentage , faster

GRAPH num_nodes: 30000  num_edges: 30000 
sampled_nodes_amount: 7653  fanout: 9 excluded edges amount: 16394 
SUBGRAPH num_nodes: 30000  num_edges: 3677
==================================================