@mszarma 2021-07-07
> did you benchmark how fast is this implementation?

Yes, i bench second version on synthetic benchmark , here are some examples of results:
OLD = remove_edges + neighbor_sampling call
NEW = neighbor_sampling with excluded edges call
==================================================
old=0.0038526 , new=0.0017021 , diff=0.0021505 seconds, diff=55.82 percentage , faster

GRAPH num_nodes: 2997  num_edges: 1909 
sampled_nodes_amount: 735  fanout: 151 excluded edges amount: 1165 
SUBGRAPH num_nodes: 2997  num_edges: 198
==================================================
==================================================
old=0.0061331 , new=0.0036194 , diff=0.0025136 seconds, diff=40.99 percentage , faster

GRAPH num_nodes: 29998  num_edges: 29797 
sampled_nodes_amount: 1126  fanout: 196 excluded edges amount: 684 
SUBGRAPH num_nodes: 29998  num_edges: 1775
==================================================
==================================================
old=0.0427086 , new=0.0160954 , diff=0.0266132 seconds, diff=62.31 percentage , faster

GRAPH num_nodes: 1019999  num_edges: 29855 
sampled_nodes_amount: 174  fanout: 3 excluded edges amount: 168 
SUBGRAPH num_nodes: 1019999  num_edges: 231 
==================================================

==================================================
old=12.5179353 , new=6.0757782 , diff=6.4421570 seconds, diff=51.46 percentage , faster

GRAPH num_nodes: 80002000  num_edges: 239989054 
sampled_nodes_amount: 689396  fanout: 6 excluded edges amount: 1990128 
SUBGRAPH num_nodes: 80002000  num_edges: 4136376
==================================================
==================================================
old=8.8966670 , new=6.6231253 , diff=2.2735417 seconds, diff=25.55 percentage , faster

GRAPH num_nodes: 80020000  num_edges: 239989189 
sampled_nodes_amount: 94441  fanout: 6 excluded edges amount: 76725 
SUBGRAPH num_nodes: 80020000  num_edges: 566646
==================================================