@zheng-da 2021-07-15
how did you benchmark it? i think you can just take one dataset (e.g., OGBN-products) and benchmark it with and without using this PR for edge sampling.