@zheng-da 2021-07-16
you need to benchmark your implementation on a large graph.