@zheng-da 2021-07-16
you can take a look at the implementation of `EdgeDataLoader`: https://doc.dgl.ai/api/python/dgl.dataloading.html#dgl.dataloading.pytorch.EdgeDataLoader
it excludes certain edges.