@BarclayII 2021-09-02
Noticed another problem: now that we are putting `exclude_eids` in `sample_neighbors`, `exclude_eids_in_frontier` may not work by itself since we need to support the GPU version as well.  Will fix this later.